submit: update put function with authorization

+++ controllers/blogs.js
@@ -17,7 +17,6 @@ blogsRouter.get('/:id', async (request, response) => {
   blog === null
     ? response.status(404).end()
     : response.json(blog)
-  // response.json(blog)
 })
 
 blogsRouter.post('/', async (request, response) => {
@@ -52,25 +51,57 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
-  await Blog.findByIdAndDelete(id)
-  // console.log('delete result', result)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(id)
+
+  if (blog) {
+    // console.log('blog', blog)
+    // console.log('logged in user', decodedToken.id)
+    // console.log('typeof blog user', typeof blog.user)
+    // console.log('typeof logged in user', typeof decodedToken.id)
+
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid, unauthorized action' })
+    }
+
+    await Blog.findByIdAndDelete(id)
+    // console.log('delete result', result)
+    response.status(204).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const id = request.params.id
   const { title, author, url, likes } = request.body
 
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
   let blog = await Blog.findById(id)
 
-  title ? blog.title = title : null
-  author ? blog.author = author : null
-  url ? blog.url = url : null
-  likes ? blog.likes = likes : null
+  if (blog) {
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid, unauthorized action' })
+    }
 
-  // send to server
-  const updatedBlog = await blog.save()
-  response.json(updatedBlog)
+    title ? blog.title = title : null
+    author ? blog.author = author : null
+    url ? blog.url = url : null
+    likes ? blog.likes = likes : null
+
+    // send to server
+    const updatedBlog = await blog.save()
+    response.json(updatedBlog)
+  }
 })
 
 module.exports = blogsRouter
